refactor: consolidate settings providers - unify into single settings_providers.dart file, delete redundant legacy files
