+++ src/client/lib/features/settings/presentation/widgets/appearance_section.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../../../gen/app_localizations.dart';
-import '../providers/settings_providers_unified.dart';
+import '../providers/settings_providers.dart';
 import 'setting_item.dart';
 import 'settings_card.dart';
