+++ src/client/lib/features/settings/presentation/widgets/performance_section.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../providers/settings_providers_unified.dart';
+import '../providers/settings_providers.dart';
 import 'setting_item.dart';
 import 'settings_card.dart';
