+++ src/client/lib/features/settings/presentation/widgets/profile_section.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../../../gen/app_localizations.dart';
-import '../providers/settings_providers_unified.dart';
+import '../providers/settings_providers.dart';
 import 'settings_card.dart';
 
 /// Profile section widget - manages user profile settings.
